fix(theme): fix tests
